+++ data/duke.Duke/tasks.txt
@@ -1,0 +1,18 @@
+T---false--- eat---null 
+D---false--- return book ---Sunday           
+D---false--- return book ---Sunday           
+E---true--- project meeting ---Mon 2-4pm           
+E---false--- project meeting ---Mon 2-4pm           
+E---true--- project meeting ---Mon 2-4pm           
+D---false--- return book ---Dec 19 2019 06:00 PM           
+E---true--- dododod ---2035-05-8 1200          
+E---false--- dododod ---May 08 2035 12:00 PM          
+E---false--- dododod ---2035-05-98 1200          
+E---false--- dododod ---Nov 01 2035 12:00 PM          
+D---true--- sleep ---Dec 12 2020 07:00 PM          
+E---false--- haha ---Dec 12 2020 12:00 AM         
+D---false--- ssss ---Dec 12 2020 12:00 AM     
+T---false--- sleep---null 
+E---false--- aaaaaaaaaaa ---2020-1-1 1100  
+E---false--- fly to Singapore---Jan 01 2020 11:00 AM  
+T---false--- die---null
