+++ src/main/java/duke/DukeExceptions.java
@@ -1,3 +1,5 @@
+package duke;
+
 /**
  * Placeholder Exception
  */
